- include seed script inside start script - exclude force:true for db.sync for production

+++ script/seed.js
@@ -15,7 +15,12 @@ const {
 
 const seed = async () => {
   try {
-    await db.sync({ force: true });
+    if (process.env.NODE_ENV === 'development') {
+      await db.sync({ force: true });
+    }
+    else {
+      await db.sync();
+    }
     console.log("seeding...");
 
     //Empty arrays so we can add instances in order to create associations between the instances later.
